% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrc_confusion_CalculateConfusion.R
\name{CalculateConfusion}
\alias{CalculateConfusion}
\title{CalculateConfuison}
\usage{
CalculateConfusion(
  model,
  signal.max = NULL,
  confusion.signal.max = signal.max,
  ...
)
}
\description{
This function calculates confusiuon waves from confusion matrix.
model need to have confusion matrix attribute
}
