% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_GenerateBootstrapSample.R
\name{GenerateBootstrapSample.FRAModel}
\alias{GenerateBootstrapSample.FRAModel}
\title{GenerateBootstrapSample.FRAModel}
\usage{
\method{GenerateBootstrapSample}{FRAModel}(
  model,
  bootstrap.number = 1,
  bootstrap.sample_size = NULL,
  parallel_cores = 1,
  ...
)
}
\arguments{
\item{model}{FRAModel}

\item{bootstrap.number}{number of bootstrap sampling}

\item{bootstrap.sample_size}{size of one bootstrap sample}
}
\description{
GenerateBootstrapSample.FRAModel
}
