% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_GenerateBootstrapSample.R
\name{GenerateBootstrapSample}
\alias{GenerateBootstrapSample}
\title{GenerateBootstrapSample}
\usage{
GenerateBootstrapSample(model, ...)
}
\arguments{
\item{model}{SCRCModel}
}
\description{
This function generate bootstrap samples
}
