% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_GenerateBootstrapSample.R
\name{GenerateBootstrapSample.SCRCModel}
\alias{GenerateBootstrapSample.SCRCModel}
\title{GenerateBootstrapSample.SCRCModel}
\usage{
\method{GenerateBootstrapSample}{SCRCModel}(
  model,
  bootstrap.number = 1,
  bootstrap.sample_size = NULL,
  parallel_cores = 1,
  ...
)
}
\arguments{
\item{model}{SCRCModel}

\item{bootstrap.number}{number of bootstrap sampling}

\item{bootstrap.sample_size}{size of one bootstrap sample}
}
\description{
GenerateBootstrapSample.SCRCModel
}
