% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_GenerateDefaultBootstrapSampleSize.R
\name{GenerateDefaultBootstrapSampleSize}
\alias{GenerateDefaultBootstrapSampleSize}
\title{GenerateDefaultBootstrapSampleSize}
\usage{
GenerateDefaultBootstrapSampleSize(model, bootstrap.sample_size, ...)
}
\arguments{
\item{model}{ITRCModel}
}
\description{
GenerateDefaultBootstrapSampleSize
}
