% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_GenerateBootstrapSample.R
\name{GenerateNoBootstrapSample.FRAModel}
\alias{GenerateNoBootstrapSample.FRAModel}
\title{GenerateNoBootstrapSample.FRAModel}
\usage{
\method{GenerateNoBootstrapSample}{FRAModel}(model, ...)
}
\arguments{
\item{model}{FRAModel}
}
\description{
This function generate bootstrap samples
}
