% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_GenerateBootstrapSample.R
\name{GenerateNoBootstrapSample}
\alias{GenerateNoBootstrapSample}
\title{GenerateNoBootstrapSample}
\usage{
GenerateNoBootstrapSample(model)
}
\description{
GenerateNoBootstrapSample
}
