% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_GenerateBootstrapSample.R
\name{GenerateNoBootstrapSample.SCRCModel}
\alias{GenerateNoBootstrapSample.SCRCModel}
\title{GenerateNoBootstrapSample.SCRCModel}
\usage{
\method{GenerateNoBootstrapSample}{SCRCModel}(model, ...)
}
\arguments{
\item{model}{SCRCModel}
}
\description{
This function generate bootstrap samples
}
