% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_Palletes.R
\name{GetLevelColors}
\alias{GetLevelColors}
\title{GetLevelColors}
\usage{
GetLevelColors(levels, levels.names = levels, ...)
}
\arguments{
\item{levels}{numeric vector}

\item{levels.names}{character vector, names of signals}

\item{...}{Arguments passed on to \code{GetPalleteColors}
\describe{
  \item{pallete}{specify colors pallete; pallete can be chosen from set
\code{c("virdis", "blue_red", "blue_green_red", "yellow_red", "orange", "bcgyr")} or as a vector of colors that will be
passed to the function \code{GetPalleteFun} to create new pallete}
  \item{pallete.args}{list, arguments to pallete (for instance to virdis function)}
}}
}
\value{
vector of colors for signal levels
}
\description{
Returns colors pallete for signals defined in \code{levels}
}
