% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_Palletes.R
\name{GetPalleteColors}
\alias{GetPalleteColors}
\title{GetPalleteColors}
\usage{
GetPalleteColors(pallete = "virdis", pallete.args = list(), ...)
}
\arguments{
\item{pallete}{specify colors pallete; pallete can be chosen from set
\code{c("virdis", "blue_red", "blue_green_red", "yellow_red", "orange", "bcgyr")} or as a vector of colors that will be
passed to the function \code{GetPalleteFun} to create new pallete}

\item{pallete.args}{list, arguments to pallete (for instance to virdis function)}
}
\value{
vector of colors pallete in RGB
}
\description{
This function return colors pallete according to internal parameters. In default return virdis pallete.
}
