% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_GetPlotTheme.R
\name{GetPlotTheme}
\alias{GetPlotTheme}
\title{GetPlotTheme}
\usage{
GetPlotTheme(theme_fun = theme_scrc, theme_fun.args = list(), ...)
}
\arguments{
\item{theme_fun}{\code{ggthemes} or \code{ggplot2} theme function. Default set to \code{theme_scrc}.}

\item{theme_fun.args}{list of parameters to \code{theme_fun}}
}
\description{
GetPlotTheme
}
