% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_GetRescaledSignalTheme.R
\name{GetRescaledSignalTheme}
\alias{GetRescaledSignalTheme}
\title{GetRescaledSignalTheme}
\usage{
GetRescaledSignalTheme(
  model = NULL,
  signals.rescale.df = NULL,
  colors = NULL,
  ...
)
}
\arguments{
\item{model}{SCRCModel object return by SCRC function}

\item{signals.rescale.df}{optional, data.frame returned from \code{rescaleSignalsValues.DataFrame}}

\item{colors}{optional, vector returned from \code{GetLevelColors}}
}
\value{
\code{list(colors = colors, signals.rescale.df = signals.rescale.df)}
}
\description{
This functions return list of objects usfeul for theme compositions that depends on signal.
If \code{signals.rescale.df} and \code{colors} are not defined, then functions \code{rescaleSignalsValues.DataFrame}
and \code{GetLevelColors} are called respectively
}
