% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_Palletes.R
\name{GetWaveColorsRescalingObject}
\alias{GetWaveColorsRescalingObject}
\title{GetWaveColorsRescalingObject}
\usage{
GetWaveColorsRescalingObject(
  pallete.fun = NULL,
  class.list,
  colors.n = 100,
  colors.min.diff = 2,
  ...
)
}
\arguments{
\item{pallete.fun}{function, declared as \code{function(n){ }} that returns n colors with increasing intensity}

\item{class.list}{numeric vector, specify levels of color sensitivity}

\item{colors.n}{numeric, number of colors in range, \code{default = 100}}

\item{colors.min.diff}{numeric, minimal difference between colors ids in colors range, \code{default = 2}}
}
\value{
function, declared as \code{function(level) {}}, that return color for level
}
\description{
Internal function that specify function which returns color for signal level.
}
