% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_ITRC.R
\name{ITRC}
\alias{ITRC}
\title{Compute Information-Theoretic Response Curve}
\usage{
ITRC(data, signal = "signal", response = "response",
  sample = "sample", bootstrap = FALSE, bootstrap.number = 2,
  bootstrap.sample_size = NULL, parallel_cores = 1, ...)
}
\arguments{
\item{data}{A data.frame or data.table object with columns \code{signal}, \code{response} and \code{sample} in a wide format}

\item{bootstrap}{\code{default = FALSE}}

\item{bootstrap.number}{\code{default = 1}}

\item{bootstrap.sample_size}{\code{default = NULL}}
}
\value{
ITRC S3 object with
}
\description{
Compute Information-Theoretic Response Curve
}
