% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SCRCmodel.R
\name{new_FRAModel}
\alias{new_FRAModel}
\title{Constructor of FRAModel S3 object}
\usage{
new_FRAModel(data, signal, response, sample)
}
\arguments{
\item{data}{A data.frame or data.table object in a wide format that describe
response (might be multidimmensional) of the samples to the signal
(now only one dimmensional). Data.frame \code{data} consists columns of names
defined by \code{sample}, \code{signal} (optional), and \code{response}.
 Each row represents a response of one sample to the input signal.
Column \code{signal} define the input signal (now only in one-dimmension).
Columns \code{response} define the multidimmensional (optional) response to
 the input signal.
Column \code{sample} specify identifaction of sample. If \code{sample} is not
 defined then sample is identified by row number.}

\item{signal}{character, specify name of the column that represents the input signal}

\item{response}{vector of chartacters, that specify names of the columns that represents
the output response}

\item{sample}{character (optional), specify name of the column that consists identifiaction of
sample}
}
\value{
FRAModel S3 object with
}
\description{
Constructor of FRAModel S3 object
}
