% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ITRCmodel.R
\name{new_ITRCModel}
\alias{new_ITRCModel}
\title{Constructor of ITRCModel S3 object}
\usage{
new_ITRCModel(data, signal, response, sample)
}
\arguments{
\item{data}{A data.frame or data.table object with columns \code{signal}, \code{response} and \code{sample} in a wide format}

\item{sample}{}
}
\value{
ITRCModel S3 object with
}
\description{
Constructor of ITRCModel S3 object
}
