% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plotConfusionMatrix.R
\name{plotCofusionMatrix}
\alias{plotCofusionMatrix}
\title{plotSCRCWaves}
\usage{
plotCofusionMatrix(
  model,
  confusion.signal = NULL,
  save.plot = FALSE,
  title_ = "",
  ylab_ = "Specific Stimulation Level",
  xlab_ = "Stimulation Level",
  ...
)
}
\arguments{
\item{model}{SCRCModel object return by SCRC function}

\item{confusion.signal}{signal for which you want to plot confusion matrix, default \code{confusion.signal == max(signal)}}

\item{title_}{character, specify title of plot, default \code{"Information Theoretic Response Curve"}}

\item{ylab_}{character, label of y axes and legend title, default \code{"Signal levels"}}

\item{xlab_}{character, label of x axes, default \code{"States number"}}
}
\description{
This functions return ggplot2 figure that visualise Information
 Theoretic Response Curves and specificity of cellular response to particular signals.
}
\details{
TODO important
}
