% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plotITRCWavesComparison.R
\name{plotFRC.Comparison}
\alias{plotFRC.Comparison}
\title{plotFRA.Comparison}
\usage{
plotFRC.Comparison(
  model,
  data,
  variable.to.compare,
  variable.to.rescale = variable.to.compare,
  theme.signal = NULL,
  ylab.right = "Cellular Response",
  data_raw_min = 0,
  theme.data.points = list(color = "black"),
  theme.data.line = list(color = "black"),
  ...
)
}
\arguments{
\item{model}{FRAModel object return by FRA function}

\item{data, }{column "model$signal" and}

\item{variable.to.compare, }{column of data}

\item{theme.signal}{optional, object returned by \code{GetRescaledSignalTheme}}

\item{ylab.right}{character, label of right y axes and legend title, default \code{"Signal levels"}}

\item{data_raw_min, }{column of data}

\item{...}{
  Arguments passed on to \code{\link[=plotFRC]{plotFRC}}
  \describe{
    \item{\code{title_}}{character, specify title of plot, default \code{"Fractional Response Curve"}}
    \item{\code{xlab_}}{character, label of x axes, default \code{"Dose"}}
    \item{\code{ylab_}}{character, label of y axes and legend title, default \code{"Cumulative fraction of cells"}}
    \item{\code{ylimits_}}{TRUE FALSE or vector of minimum and maximum of y axes}
    \item{\code{fill.guide_}}{logical, specify if legend should be displayed}
    \item{\code{plot.heterogeneity}}{logical, specify if heterogeneity should be added to FRC plot, default \code{TRUE}}
  }}

\item{signal.max}{...}
}
\description{
This functions return ggplot2 figure that visualise fractional response curve (FRC) compared to standard response curve
}
\details{
TODO important
}
