% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plotConfusionMatrix.R
\name{plotHeterogeneityPieCharts}
\alias{plotHeterogeneityPieCharts}
\title{plotHeterogeneityPieCharts}
\usage{
plotHeterogeneityPieCharts(
  model,
  max.signal = NULL,
  save.plot = FALSE,
  title_ = "Cell-to-cel heterogeneity",
  ylab_ = "dose",
  xlab_ = "dose for which response is typical",
  ...
)
}
\arguments{
\item{model}{FRAModel object return by FRA function}

\item{max.signal}{maximal signal for which the cell-to-cell heterogeneity is plotted, default \code{confusion.signal == max(signal)}}

\item{title_}{character, specify title of plot, default \code{"Cell-to-cel heterogeneity"}}

\item{ylab_}{character, label of y axes, default \code{"dose"}}

\item{xlab_}{character, label of x axes, default \code{"dose for which response is typical"}}
}
\description{
This functions return ggplot2 figure that visualise using pie charts what fraction of cells
 exposed to one dose exhibits responses in the range characteristic for other doses.
}
