% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plotITRCWavesComparison.R
\name{plotITRCWaves.Comparison}
\alias{plotITRCWaves.Comparison}
\title{plotITRCWaves.Comparison}
\usage{
plotITRCWaves.Comparison(data, variable.to.compare, ...)
}
\arguments{
\item{data, }{column "model$signal" and}

\item{variable.to.compare, }{column of data}

\item{...}{Arguments passed on to \code{plotITRCWaves}
\describe{
  \item{model}{ITRCModel object return by ITRC function}
  \item{title_}{character, specify title of plot, default \code{"Information Theoretic Response Curve"}}
  \item{xlab_}{character, label of x axes, default \code{"States number"}}
  \item{ylab_}{character, label of y axes and legend title, default \code{"Signal levels"}}
  \item{ylimits_}{TRUE FALSE or vector of minimum and maximum of y axes}
  \item{fill.guide_}{logical, specify if legend should be displayed}
}}
}
\description{
This functions return ggplot2 figure that visualise Information
 Theoretic Response Curves and specificity of cellular response to particular signals.
}
\details{
TODO important
}
