% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plotITRCWaves.R
\name{plotSCRCWaves}
\alias{plotSCRCWaves}
\title{plotSCRCWaves}
\usage{
plotSCRCWaves(
  model,
  title_ = "Information Theoretic Response Curve",
  xlab_ = "States number",
  ylab_ = "Signal levels",
  fill.guide_ = "legend",
  ylimits_ = TRUE,
  alpha_ = 0.5,
  getScaleY = getScaleY.SCRC,
  theme.signal = NULL,
  ...
)
}
\arguments{
\item{model}{SCRCModel object return by SCRC function}

\item{title_}{character, specify title of plot, default \code{"Information Theoretic Response Curve"}}

\item{xlab_}{character, label of x axes, default \code{"States number"}}

\item{ylab_}{character, label of y axes and legend title, default \code{"Signal levels"}}

\item{fill.guide_}{logical, specify if legend should be displayed}

\item{ylimits_}{TRUE FALSE or vector of minimum and maximum of y axes}

\item{theme.signal}{optional, object returned by \code{GetRescaledSignalTheme}}

\item{...}{
  Arguments passed on to \code{\link[=rescaleSignalsValues]{rescaleSignalsValues}}
  \describe{
    \item{\code{rescale.fun}}{parameter, that defines a function used for rescaling signals in plots.
There are three built-in functions, that can be chosen:
(1) \code{'factor'} - signals treated as factors (default) with levels defined in list \code{rescale.fun.args},
(2) \code{'numeric'},
(3) \code{logarithmic} - with base defined in \code{rescale.fun.args} - default: \code{e = exp(1)}.
Function must be defined as a lambda construct \code{function(x, ...){...}}.}
    \item{\code{rescale.fun.args}}{list of the arguments to defaults \code{rescale.fun}}
  }}
}
\description{
This functions return ggplot2 figure that visualise Information
 Theoretic Response Curves and specificity of cellular response to particular signals.
}
\details{
TODO important
}
