% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleResponse.R
\name{rescaleDataToITRC.DataFrame}
\alias{rescaleDataToITRC.DataFrame}
\title{rescaleDataToITRC.DataFrame}
\usage{
rescaleDataToITRC.DataFrame(model, data = NULL, variable_ = NULL,
  variable.rescaled_ = NULL, data_raw_min = NULL, a = NULL,
  b = NULL, ...)
}
\arguments{
\item{model}{ITRCModel}

\item{data}{data to rescale}

\item{variable_}{variable_ TODO}

\item{variable.rescaled_}{variable.rescaled_ TODO}

\item{...}{Arguments passed on to \code{rescaleDataToITRC.Params}
\describe{
  \item{model}{ITRCModel}
  \item{data}{data to rescale}
  \item{variable_}{variable_ TODO}
  \item{data_raw_min}{data_raw_min 0 or min of data}
}}
}
\description{
rescaleDataToITRC.DataFrame
}
