% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleResponse.R
\name{rescaleDataToITRC.Params}
\alias{rescaleDataToITRC.Params}
\title{rescaleDataToITRC.Params}
\usage{
rescaleDataToITRC.Params(model = NULL, data, variable.to.compare,
  data_raw_min = NULL, ...)
}
\arguments{
\item{model}{ITRCModel}

\item{data}{data to rescale}

\item{variable.to.compare}{variable.to.compare TODO}

\item{data_raw_min}{data_raw_min 0 or min of data}
}
\description{
rescaleDataToITRC.Params
}
