% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleResponse.R
\name{rescaleDataToSCRC.DataFrame}
\alias{rescaleDataToSCRC.DataFrame}
\title{rescaleDataToSCRC.DataFrame}
\usage{
rescaleDataToSCRC.DataFrame(
  model,
  data = NULL,
  variable.to.compare = NULL,
  variable.rescaled = NULL,
  variable.to.rescale = NULL,
  data_raw_min = NULL,
  a = NULL,
  b = NULL,
  ...
)
}
\arguments{
\item{model}{SCRCModel}

\item{data}{data to rescale}

\item{variable.to.compare}{variable.to.compare TODO}

\item{variable.rescaled}{variable.rescaled TODO}

\item{...}{
  Arguments passed on to \code{\link[=rescaleDataToSCRC.Params]{rescaleDataToSCRC.Params}}
  \describe{
    \item{\code{data_raw_min}}{data_raw_min 0 or min of data}
  }}
}
\description{
rescaleDataToSCRC.DataFrame
}
