% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleResponse.R
\name{rescaleDataToSCRC.Params}
\alias{rescaleDataToSCRC.Params}
\title{rescaleDataToSCRC.Params}
\usage{
rescaleDataToSCRC.Params(
  model = NULL,
  data,
  variable.to.compare,
  data_raw_min = NULL,
  signal.max = NULL,
  ...
)
}
\arguments{
\item{model}{SCRCModel}

\item{data}{data to rescale}

\item{variable.to.compare}{variable.to.compare TODO}

\item{data_raw_min}{data_raw_min 0 or min of data}
}
\description{
rescaleDataToSCRC.Params
}
