% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleSignal.R
\name{rescaleSignalsValues.factor}
\alias{rescaleSignalsValues.factor}
\title{rescaleSignalsValues.factor}
\usage{
rescaleSignalsValues.factor(levels, ...)
}
\description{
rescaleSignalsValues.factor
}
