% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleSignal.R
\name{rescaleSignalsValues.numeric}
\alias{rescaleSignalsValues.numeric}
\title{rescaleSignalsValues.numeric}
\usage{
rescaleSignalsValues.numeric(...)
}
\description{
rescaleSignalsValues.numeric
}
