% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_ThemeITRC.R
\name{theme_itrc}
\alias{theme_itrc}
\title{theme_itrc}
\usage{
theme_itrc(theme.base_size = 12, theme.base_family = "sans",
  theme.title_size = 36, theme.text_size = 3 * theme.title_size/4,
  theme.margins = c(1, 1, 1, 1), legend.position = "right", ...)
}
\arguments{
\item{theme.base_size}{= 12,}

\item{theme.base_family}{= "sans",}

\item{theme.title_size}{= 36,}

\item{theme.text_size}{= 3*theme.title_size/4,}

\item{theme.margins}{= c(1,1, 1, 1),}

\item{legend.position}{= "right",}
}
\value{
ggthemes::theme_foundation object
}
\description{
This function specify theme used in the ITRC package
}
