% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_GenerateBootstrapSample.R
\name{GenerateNoBootstrapSample.ITRCModel}
\alias{GenerateNoBootstrapSample.ITRCModel}
\title{GenerateNoBootstrapSample.ITRCModel}
\usage{
\method{GenerateNoBootstrapSample}{ITRCModel}(model, ...)
}
\description{
GenerateNoBootstrapSample.ITRCModel
}
