% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_ITRC.R
\name{ITRC}
\alias{ITRC}
\title{Compute Information-Theoretic Response Curve}
\usage{
ITRC(data, signal = "signal", response = "response",
  sample = "sample", bootstrap = FALSE, bootstrap.number = 2,
  bootstrap.sample_size = NULL, parallel_cores = 1, ...)
}
\arguments{
\item{data}{A data.frame or data.table object in a wide format that describe
response (might be multidimmensional) of the samples to the signal
(now only one dimmensional). Data.frame \code{data} consists columns of names
defined by \code{sample}, \code{signal} (optional), and \code{response}.
 Each row represents a response of one sample to the input signal.
Column \code{signal} define the input signal (now only in one-dimmension).
Columns \code{response} define the multidimmensional (optional) response to
 the input signal.
Column \code{sample} specify identifaction of sample. If \code{sample} is not
 defined then sample is identified by row number.}

\item{signal}{character, specify name of the column that represents the input signal}

\item{response}{vector of chartacters, that specify names of the columns that represents
the output response}

\item{sample}{character (optional), specify name of the column that consists identifiaction of
sample}

\item{bootstrap}{\code{default = FALSE}}

\item{bootstrap.number}{\code{default = 1}}

\item{bootstrap.sample_size}{\code{default = NULL}}
}
\value{
ITRC S3 object with
}
\description{
This function calculate Information Theroetic Response Curves.
}
