% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plotITRCWavesComparison.R
\name{plotITRCWaves.Comparison}
\alias{plotITRCWaves.Comparison}
\title{plotITRCWaves.Comparison}
\usage{
plotITRCWaves.Comparison(model, data, variable.to.compare,
  variable.to.rescale = variable.to.compare, theme.signal = NULL,
  ylab.right = "Cellular Response", data_raw_min = 0,
  theme.data.points = list(color = "black"),
  theme.data.line = list(color = "black"), ...)
}
\arguments{
\item{model}{ITRCModel object return by ITRC function}

\item{data, }{column "model$signal" and}

\item{variable.to.compare, }{column of data}

\item{theme.signal}{optional, object returned by \code{GetRescaledSignalTheme}}

\item{ylab.right}{character, label of right y axes and legend title, default \code{"Signal levels"}}

\item{data_raw_min, }{column of data}

\item{...}{Arguments passed on to \code{plotITRCWaves}
\describe{
  \item{model}{ITRCModel object return by ITRC function}
  \item{title_}{character, specify title of plot, default \code{"Information Theoretic Response Curve"}}
  \item{xlab_}{character, label of x axes, default \code{"States number"}}
  \item{ylab_}{character, label of y axes and legend title, default \code{"Signal levels"}}
  \item{ylimits_}{TRUE FALSE or vector of minimum and maximum of y axes}
  \item{fill.guide_}{logical, specify if legend should be displayed}
  \item{theme.signal}{optional, object returned by \code{GetRescaledSignalTheme}}
}}
}
\description{
This functions return ggplot2 figure that visualise Information
 Theoretic Response Curves and specificity of cellular response to particular signals.
}
\details{
TODO important
}
